Change cosine value from integer to float field

+++ jobs/models.py
@@ -51,7 +51,7 @@ class Job(models.Model):
     skills = models.CharField(max_length=255, blank=True, null=True)
     salary = models.CharField(max_length=20)
     description = RichTextField()
-    cosinevalue = models.IntegerField(null=True)
+    cosinevalue = models.FloatField(null=True)
 
     created_date = models.DateField(null=True, blank=True)
